+++ src/builtins/argparse.rs
@@ -60,6 +60,7 @@ enum UnknownHandling {
 #[derive(Default)]
 struct ArgParseCmdOpts<'args> {
     unknown_handling: UnknownHandling,
+    unknown_arguments: ArgCardinality,
     print_help: bool,
     stop_nonopt: bool,
     min_args: usize,
@@ -78,16 +79,18 @@ impl ArgParseCmdOpts<'_> {
     fn new() -> Self {
         Self {
             max_args: usize::MAX,
+            unknown_arguments: ArgCardinality::Optional,
             ..Default::default()
         }
     }
 }
 
-const SHORT_OPTIONS: &wstr = L!("+:hn:simx:N:X:");
+const SHORT_OPTIONS: &wstr = L!("+:hn:sima:x:N:X:");
 const LONG_OPTIONS: &[WOption] = &[
     wopt(L!("stop-nonopt"), ArgType::NoArgument, 's'),
     wopt(L!("ignore-unknown"), ArgType::NoArgument, 'i'),
     wopt(L!("move-unknown"), ArgType::NoArgument, 'm'),
+    wopt(L!("unknown-arguments"), ArgType::RequiredArgument, 'a'),
     wopt(L!("name"), ArgType::RequiredArgument, 'n'),
     wopt(L!("exclusive"), ArgType::RequiredArgument, 'x'),
     wopt(L!("help"), ArgType::NoArgument, 'h'),
@@ -531,7 +534,26 @@ fn parse_cmd_opts<'args>(
                 } else {
                     UnknownHandling::Move
                 }
-            },
+            }
+            'a' => {
+                let kind = w.woptarg.unwrap();
+                opts.unknown_arguments = if kind == L!("optional") {
+                    ArgCardinality::Optional
+                } else if kind == L!("required") {
+                    ArgCardinality::Once
+                } else if kind == L!("none") {
+                    ArgCardinality::None
+                } else {
+                    // No need for a way to set ArgCardinality::AtLeastOnce, as that only differs from ArgCardinality::Once
+                    // in the way _flag_ variables are set, which are not set for unknown options.
+                    streams.err.append(wgettext_fmt!(
+                        "%ls: Invalid --unknown-arguments value '%ls'\n",
+                        cmd,
+                        kind
+                    ));
+                    return Err(STATUS_INVALID_ARGS);
+                }
+            }
             // Just save the raw string here. Later, when we have all the short and long flag
             // definitions we'll parse these strings into a more useful data structure.
             'x' => opts.raw_exclusive_flags.push(w.woptarg.unwrap()),
@@ -862,7 +884,7 @@ fn argparse_parse_flags<'args>(
     let mut w = WGetopter::new(&short_options, &long_options, args);
     while let Some((opt, longopt_idx)) = w.next_opt_indexed() {
         let is_long_flag = longopt_idx.is_some();
-        let retval: BuiltinResult = match opt {
+        match opt {
             ':' => {
                 builtin_missing_argument(
                     parser,
@@ -871,7 +893,7 @@ fn argparse_parse_flags<'args>(
                     args_read[w.wopt_index - 1],
                     false,
                 );
-                Err(STATUS_INVALID_ARGS)
+                return Err(STATUS_INVALID_ARGS);
             }
             '?' => {
                 // It's not a recognized flag. See if it's an implicit int flag.
@@ -885,24 +907,44 @@ fn argparse_parse_flags<'args>(
                         &mut w,
                         is_long_flag,
                         streams,
-                    )
+                    )?;
                 } else if opts.unknown_handling == UnknownHandling::Error {
                     streams.err.append(wgettext_fmt!(
                         BUILTIN_ERR_UNKNOWN,
                         opts.name,
                         args_read[w.wopt_index - 1]
                     ));
-                    Err(STATUS_INVALID_ARGS)
+                    return Err(STATUS_INVALID_ARGS);
                 } else {
                     debug_assert!(!is_long_flag); // The option is unknown, so there's no long opt index it could have used
                     let is_long_flag = arg_contents.starts_with("-"); // arg_contents already skipped over the first '-'
+
+                    if is_long_flag {
+                        // For some reason, wgetopt parses unknown long options as if they where a short '-' flag,
+                        // followed by the long flag name, interpreted either as the value for '-' or as remaining short options
+                        // so this fixes that
+                        let rest = if let Some(value) = w.woptarg {
+                            debug_assert!(w.remaining_text.is_empty());
+                            value
+                        } else {
+                            w.remaining_text
+                        };
+                        // The arguments was of form --<long-flag>=<value>, so extract out <value>
+                        if let Some(i) = rest.find_char('=') {
+                            w.woptarg = Some(&rest[i + 1..]);
+                        }
+                        // Ensure w.remaining_text is not misinterpreted as the value of the flag, or as remaining short options
+                        w.remaining_text = L!("");
+                    }
 
                     // Any unrecognized option is put back if ignore_unknown is used.
                     // This allows reusing the same argv in multiple argparse calls,
                     // or just ignoring the error (e.g. in completions).
 
-                    // First we consume any remaining text as if it was the option's argument
-                    if !w.remaining_text.is_empty() {
+                    // First we consume any remaining text as if it was the options argument
+                    if opts.unknown_arguments != ArgCardinality::None
+                        && !w.remaining_text.is_empty()
+                    {
                         debug_assert!(w.woptarg.is_none()); // Both don't make sense
                         w.woptarg = Some(w.remaining_text);
                         // Explain to wgetopt that we want to skip to the next arg,
@@ -910,21 +952,68 @@ fn argparse_parse_flags<'args>(
                         w.remaining_text = L!("");
                     }
 
+                    // If unknown options require arguments, but there weren't any worked out above,
+                    // take the next element of w.argv as the argument
+                    let separate_value =
+                        if opts.unknown_arguments == ArgCardinality::Once && w.woptarg.is_none() {
+                            if w.wopt_index < w.argv.len() {
+                                w.wopt_index += 1; // Tell wgetop to skip over the options value
+                                Some(w.argv[w.wopt_index - 1])
+                            } else {
+                                // the option is at the end of argv, so it has no argument
+                                streams.err.append(wgettext_fmt!(
+                                    BUILTIN_ERR_MISSING,
+                                    opts.name,
+                                    args_read[w.wopt_index - 1]
+                                ));
+                                return Err(STATUS_INVALID_ARGS);
+                            }
+                        } else {
+                            None
+                        };
+
+                    // If the option uses the long flag syntax with a value (i.e. --<flag-name>=<value>)
+                    if opts.unknown_arguments == ArgCardinality::None
+                        && is_long_flag
+                        && arg_contents.contains('=')
+                    {
+                        // Repport an error that says the option is unknown, which is what argparse and all the other commands report
+                        // when you use --<flag-name>=<value> for a flag that doesn't take an argument
+                        streams.err.append(wgettext_fmt!(
+                            BUILTIN_ERR_UNKNOWN,
+                            opts.name,
+                            args_read[w.wopt_index - 1]
+                        ));
+                        return Err(STATUS_INVALID_ARGS);
+                    }
+
                     if opts.unknown_handling == UnknownHandling::Ignore {
                         // Now by calling delete_flag we ensure that if the unknown flag is precceded by known flags, the known flags are kept in $argv_opts, and not added to $argv.
                         // (any argument to the option is also returned in unknown_flag, and removed from opts.argv_opts)
                         let unknown_flag = delete_flag(&mut w, is_long_flag);
                         opts.args.push(unknown_flag);
+                        // If the value of the argument was the next element of the input arguments, record it
+                        // to be stored in the new $argv value
+                        if let Some(value) = separate_value {
+                            opts.args.push(Cow::Borrowed(value));
+                        }
                     } else {
                         assert!(opts.unknown_handling == UnknownHandling::Move);
-                        // Nothing more to do, w.argv_opts will already contain the option, and its value (if any)
+                        // w.argv_opts will already contain the option and its value, unless the value was given as a seperate argument
+                        if let Some(value) = separate_value {
+                            w.argv_opts.push(Cow::Borrowed(value));
+                        }
+                    }
+
+                    // Make wgetopt keep reading this argument for more options
+                    if !w.remaining_text.is_empty() {
+                        w.wopt_index -= 1;
                     }
 
                     // Work around weirdness with wgetopt, which crashes if we `continue` here.
                     if w.wopt_index == argc {
                         break;
                     }
-                    Ok(SUCCESS)
                 }
             }
             NON_OPTION_CHAR => {
@@ -937,9 +1026,10 @@ fn argparse_parse_flags<'args>(
                 continue;
             }
             // It's a recognized flag.
-            _ => handle_flag(parser, opts, opt, is_long_flag, &mut w, streams),
-        };
-        retval?;
+            _ => {
+                handle_flag(parser, opts, opt, is_long_flag, &mut w, streams)?;
+            }
+        }
     }
     opts.args_opts = w.argv_opts;